BatSense: change from document-script injection to 99% undetectable content-script. Also (and consequently) gain module and code re-use.
