+++ components/brave_extension/extension/brave_extension/batsense/content/creativeFetch/custom-event-send.js
@@ -1,0 +1,11 @@
+
+export default function sendCustomEvent (batAd) {
+  window.dispatchEvent(new CustomEvent('bat-ad-position-ready', {
+    detail: {
+      id: batAd.batAdId,
+      isResponsive: batAd.isResponsive,
+      sizes: batAd.sizes
+    },
+    bubbles: true
+  }))
+}
