+++ components/brave_extension/extension/brave_extension/batsense/content/pageLifecycle/run-on-loaded.js
@@ -1,0 +1,7 @@
+export default function runOnPageLoaded (fn) {
+  if (document.readyState === 'loading') {
+    document.addEventListener('DOMContentLoaded', fn)
+  } else {
+    fn()
+  }
+}
